requirements.txt
pandas==2.2.3
requests==2.32.3
sqlalchemy==2.0.36
python-dotenv==1.0.1